Removes class/struct definition nodes and instead modifies the corresponding object nodes.
