+++ src/nodes/ClassDefinitionObject.hpp
@@ -27,14 +27,20 @@ public:
      * pass vectors to the variables and methods we actually own. The others
      * will be in a parent class if provided.
      */
-    ClassDefinitionObject(std::vector<AddVariableNode *> variableDefs_,
-                          std::vector<FunctionNode *> methodDefs_,
-                          ClassDefinitionObject *parent_ = nullptr);
+    ClassDefinitionObject(std::string typeName_,
+                          std::string parentTypeName_,
+                          std::vector<AddVariableNode *> variableDefs_,
+                          std::vector<FunctionNode *> methodDefs_);
 
     /**
      * Destructor deletes all method definition nodes.
      */
     ~ClassDefinitionObject() override;
+
+    /**
+     * Install object in current scope.
+     */
+    BaseObject *evaluate(Scope &scope) override;
 
     /**
      * Calls evaluate method on all method nodes. Installs them in argument scope.
@@ -45,7 +51,7 @@ protected:
     /**
      * Builds the method map from parent classes.
      */
-    void buildMethodDefsHashMap();
+    void buildMethodDefsHashMap(const Scope &scope);
 
     /**
      * Stores FunctionNodes. We will call evaluate() method to add them to class scope.
