+++ src/nodes/Nodes.hpp
@@ -18,7 +18,7 @@
 #include "AssignNode.hpp"
 #include "BinaryNode.hpp"
 #include "BreakNode.hpp"
-#include "ClassDefinitionNode.hpp"
+#include "ClassDefinitionObject.hpp"
 #include "ClassNode.hpp"
 #include "DoWhileNode.hpp"
 #include "FileNode.hpp"
@@ -33,6 +33,6 @@
 #include "PrefixIncrementNode.hpp"
 #include "ProgramNode.hpp"
 #include "ReturnNode.hpp"
-#include "StructDefinitionNode.hpp"
+#include "StructDefinitionObject.hpp"
 #include "StructNode.hpp"
 #include "WhileNode.hpp"
