+++ src/nodes/StructDefinitionObject.hpp
@@ -9,6 +9,7 @@
 
 #pragma once
 #include "AddVariableNode.hpp"
+#include "BaseNode.hpp"
 #include "BaseObject.hpp"
 #include <unordered_map>
 #include <unordered_set>
@@ -19,15 +20,16 @@
  * will be stored in the scope along with the struct name. We can then use this
  * to construct struct instances.
  */
-class StructDefinitionObject : public BaseObject
+class StructDefinitionObject : public BaseObject, public BaseNode
 {
 public:
     /**
      * Supply a vector of nodes for constructing the struct. This class will take
      * ownership and free these later.
      */
-    StructDefinitionObject(std::vector<AddVariableNode *> variableDefs_,
-                           StructDefinitionObject *parent_ = nullptr);
+    StructDefinitionObject(std::string typeName_,
+                           std::string parentTypeName_,
+                           std::vector<AddVariableNode *> variableDefs_);
 
     /**
      * Destructor deletes all nodes in variable definitions.
@@ -35,8 +37,15 @@ public:
     ~StructDefinitionObject() override;
 
     /**
+     * Registers this class in the current scope with name. Ownership will pass
+     * to the scope. Careful! Going this route means we don't have to create a
+     * separate Node to create an Object.
+     */
+    BaseObject *evaluate(Scope &scope) override;
+
+    /**
      * No destructor provided. Should not be possible to copy the struct definition
-     * as you would expect. If this were to be implemented, all nodes we area
+     * as you would expect. If this were to be implemented, all nodes we are
      * storing would have to be copied.
      */
     StructDefinitionObject *clone() const final
@@ -53,9 +62,14 @@ public:
 protected:
     /**
      * Builds the variableDefsMap from any parent classes and checks for clashing
-     * variables
+     * variables.
      */
-    void buildVariableDefHashMap();
+    void buildVariableDefHashMap(const Scope &scope);
+
+    /**
+     * Returns a pointer to the parent struct or nullptr if not found.
+     */
+    StructDefinitionObject *lookupParent(const Scope &scope) const;
 
     /**
      * Store AddVariableNode so class can create all defined variables. We take
@@ -71,7 +85,20 @@ protected:
     std::unordered_map<std::string, AddVariableNode *> allVariableDefsMap;
 
     /**
-     * Parent struct we inherit from.
+     * Type name for struct.
      */
-    StructDefinitionObject *parent{nullptr};
+    std::string typeName;
+
+    /**
+     * Type name for parent.
+     */
+    std::string parentTypeName;
+
+    /**
+     * We activate the definition once evaluate is called. This is when we can
+     * try to locate the definition of the parent if there is one and install
+     * it in the scope. We also refuse to allow more than one definition in the
+     * scope as you would expect so evaluate() cannot be called more than once!
+     */
+    bool active{false};
 };
